requirement.txt
asgiref==3.4.1
attrs==21.4.0
az-iranian-bank-gateways==1.6.15
backports.zoneinfo==0.2.1
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.10
Deprecated==1.2.13
Django==4.0
django-extensions==3.1.5
django-js-asset==1.2.2
django-mptt==0.13.4
django-redis==5.2.0
django-user-agents==0.4.0
djangorestframework==3.13.1
idna==3.3
isodate==0.6.1
lxml==4.7.1
packaging==21.3
Pillow==8.4.0
platformdirs==2.4.1
psycopg2-binary==2.9.2
pycryptodome==3.12.0
pygraphviz==1.7
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
redis==4.1.0
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
ua-parser==0.10.0
urllib3==1.26.8
user-agents==2.2.0
wrapt==1.13.3
zeep==4.1.0
